Move array class to math
